backups/refactoring_20250904_003648/requirements/base.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Dependencies Injection
dependency-injector==4.41.0

# Data & Analysis
pandas==2.1.4
numpy==1.25.2
ta==0.10.2

# External APIs
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0
PyGithub==1.59.1
openai==1.3.8
alpha-vantage==2.3.1
yfinance==0.2.28

# Async Support
asyncio-mqtt==0.16.1

# Caching
redis==5.0.1

# Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Validation
marshmallow==3.20.1
pydantic==2.5.0

# Environment
python-dotenv==1.0.0

# Discord Integration
discord-webhook==1.3.0

# Job Scheduling
APScheduler==3.10.4

# Cryptography
cryptography==41.0.7

# CLI Framework
click==8.1.7
rich==13.6.0
typer==0.9.0
